requirements.txt
flask==2.0.1
torch==2.0.0
torchvision==0.15.0
Pillow==9.5.0
clip @ git+https://github.com/openai/CLIP.git
numpy==1.24.3
scikit-learn==1.2.2